% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{acml.lmem}
\alias{acml.lmem}
\title{Fitting function: ACML for a linear mixed effects model (random intercept and slope)}
\usage{
acml.lmem(y, x, z, id, w.function = "mean", InitVals, cutpoints = c(0, 5),
  SampProb = c(1, 1, 1), SampProbi = rep(1, length(y)), ProfileCol = NA)
}
\arguments{
\item{y}{response vector}

\item{x}{sum(n_i) by p design matrix for fixed effects}

\item{z}{sum(n_i) by 2 design matric for random effects (intercept and slope)}

\item{id}{sum(n_i) vector of subject ids}

\item{w.function}{options include "mean" "intercept" "slope" and "bivar"}

\item{InitVals}{starting values for c(beta, log(sigma0), log(sigma1), rho, log(sigmae))}

\item{cutpoints}{cutpoints defining the sampling regions [bivariate Q_i: a vector of length 4 c(xlow, xhigh, ylow, yhigh);
univariate Q_i: a vector of length K c(k1,k2, ... K) to define the cutpoints for Q_i based sampling regions]}

\item{SampProb}{Sampling probabilities from within each region [bivariate Q_i: a vector of length 2 c(central region, outlying region);
univariate Q_i: a vector of length K+1 with sampling probabilities for each region]}

\item{SampProbi}{Subject specific sampling probabilities.  A vector of length sum(n_i).  Not used unless using weighted Likelihood}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters
while fixing these columns at the values of params[ProfileCol]}

\item{param}{parameter vector c(beta, log(sigma0), log(sigma1), rho, sigmae)}
}
\value{
Ascertainment corrected Maximum likelihood: Ests, covar, LogL, code, robcov
}
\description{
Fitting function: ACML for a linear mixed effects model (random intercept and slope)
}

