% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions5.R
\name{acml.lmem2}
\alias{acml.lmem2}
\title{Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)}
\usage{
acml.lmem2(formula.fixed, formula.random, data, id, w.function = "mean",
  InitVals, cutpoints = c(0, 5), SampProb = NA, Weights = NA,
  ProfileCol = NA)
}
\arguments{
\item{formula.fixed}{formula for the fixed effects (of the form y~x)}

\item{formula.random}{formula for the random effects (of the form ~z).  Right now this model only fits random intercept and slope models.}

\item{data}{data frame that should contain everything in formula.fixed, formula.random, id, and Weights.  It does not include: w.function, cutpoints, SampProb}

\item{id}{sum(n_i) vector of subject ids (a variable contained in data)}

\item{w.function}{A sum(n_i) vector with possible values that include "mean" "intercept" "slope" and "bivar."  Do not use "bivar" as it has not been tested.  There should be one unique value per subject but n_i replicates of that value.  Note that w.function should NOT be in the dat dataframe.}

\item{InitVals}{starting values for c(beta, log(sigma0), log(sigma1), log((1+rho)/(1-rho)), log(sigmae))}

\item{cutpoints}{A matrix with the first dimension equal to sum(n_i).  These cutpoints define the sampling regions for individual subjects.  If using a low, medium, high, sampling scheme, this is a sum(n_i) by 2 matrix that must be a distinct object not contained in the dat dataframe.  Each row is a vector of length 2 c(k1,k2) to define the sampling regions, i.e., low, middle, high.  Each subject should have n_i rows of the same values.}

\item{SampProb}{A matrix with the first dimension equal to sum(n_i).   Sampling probabilities from within each region. Each row is a vector of length 3 with sampling probabilities for each region. Each subject should have n_i rows of the same values.  Not in data.}

\item{Weights}{Subject specific sampling weights.  A vector of length sum(n_i).  This should be a variable in the data dataframe. It should only be used if doing IPWL.  Note if doing IPWL, only use robcov (robust variances) and not covar.  If not doing IPWL, this must be a vectors of 1s.}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maximizing the log likelihood for all other parameters
while fixing these columns at the values of InitVals[ProfileCol]}
}
\value{
Ascertainment corrected Maximum likelihood: Ests, covar, LogL, code, robcov
}
\description{
Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)
}
