% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions4.R
\name{logACi1q.score2}
\alias{logACi1q.score2}
\title{Gradient of the log of the ascertainment correction piece for sampling based on univariate Q_i}
\usage{
logACi1q.score2(subjectData, beta, sigma0, sigma1, rho, sigmae)
}
\arguments{
\item{subjectData}{a list containing: yi, xi, zi, Weights.i, w.function.i, SampProb.i, cutpoints.i}

\item{beta}{mean model parameter p-vector}

\item{sigma0}{std dev of the random intercept distribution}

\item{sigma1}{std dev of the random slope distribution}

\item{rho}{correlation between the random intercept and slope}

\item{sigmae}{std dev of the measurement error distribution}
}
\value{
gradient of the log transformed ascertainment correction under univariate $Q_i$
}
\description{
Calculate the gradient of the log transformed ascertainment correction for sampling based on univariate Q_i
}
