% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{logACi2q.score}
\alias{logACi2q.score}
\title{Gradient of the log of the ascertainment correction piece for sampling based on bivariate Q_i}
\usage{
logACi2q.score(subjectData, w.function, beta, sigma0, sigma1, rho, sigmae,
  cutpoints, SampProb)
}
\arguments{
\item{subjectData}{a list containing: yi, xi, zi}

\item{w.function}{Sampling variable q_i function "mean", "intercept", "slope", "bivar".  This only gets called if w.function="bivar"}

\item{beta}{mean model parameter p-vector}

\item{sigma0}{std dev of the random intercept distribution}

\item{sigma1}{std dev of the random slope distribution}

\item{rho}{correlation between the random intercept and slope}

\item{sigmae}{std dev of the measurement error distribution}

\item{cutpoints}{cutpoints defining the sampling regions. (a vector of length 4 c(xlow, xhigh, ylow, yhigh))}

\item{SampProb}{Sampling probabilities from within each region (vector of length 2 c(central region, outlying region)).}
}
\value{
gradient of the log transformed ascertainment correction under the bivariate sampling design
}
\description{
Calculate the gradient of the log transformed ascertainment correction under designs that sample based on a bivariate Q_i (numerically)
}

