% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions5.R
\name{logACi2q.score2}
\alias{logACi2q.score2}
\title{Gradient of the log of the ascertainment correction piece for sampling based on bivariate Q_i.}
\usage{
logACi2q.score2(subjectData, beta, sigma.vc, rho.vc, sigma.e)
}
\arguments{
\item{subjectData}{a list containing: yi, xi, zi, Weights.i, w.function, SampProb, cutpoints}

\item{beta}{mean model parameter p-vector}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}

\item{sigmae}{std dev of the measurement error distribution}
}
\value{
gradient of the log transformed ascertainment correction under the bivariate sampling design
}
\description{
Calculate the gradient of the log transformed ascertainment correction under designs that sample based on a bivariate Q_i (numerically). Note that I used mu_q[1:2] and sigma_q[1:2,1:2] in the pmvnorm call to force bivariate intercept and slope sampling for the first outcome.  This will need to be adapted when we want to sample based on, say, the two slopes for the first and second outcomes
}
