% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{LogLikeiC}
\alias{LogLikeiC}
\title{Calculate the ss contributions to the conditional likelihood for the univariate and bivariate sampling cases.}
\usage{
LogLikeiC(subjectData, w.function, beta, sigma0, sigma1, rho, sigmae, cutpoints,
  SampProb)
}
\arguments{
\item{subjectData}{a list containing: yi, xi, zi, SampProbi.i}

\item{w.function}{options include "mean" "intercept" "slope" and "bivar"}

\item{beta}{mean model parameter p-vector}

\item{sigma0}{std dev of the random intercept distribution}

\item{sigma1}{std dev of the random slope distribution}

\item{rho}{correlation between the random intercept and slope}

\item{sigmae}{std dev of the measurement error distribution}

\item{cutpoints}{cutpoints defining the sampling regions. (a vector of length 4 c(xlow, xhigh, ylow, yhigh))}

\item{SampProb}{Sampling probabilities from within each region (vector of length 2 c(central region, outlying region)).}
}
\value{
ss contributions to the conditional log likelihood.  This is an internal function used by LogLikeC
}
\description{
Calculate the ss contributions to the conditional likelihood for the univariate and bivariate sampling cases.
}

