% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{odsBiv}
\alias{odsBiv}
\title{Bivariate Test Data}
\format{An object of class \code{data.frame} with 2520 rows and 7 columns.}
\usage{
odsBiv
}
\description{
XXX Describe derivation of data set here and what the columns mean
}
\keyword{data}
