% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{odsInt}
\alias{odsInt}
\title{Intercept Test Data}
\format{An object of class \code{data.frame} with 2670 rows and 8 columns.}
\usage{
odsInt
}
\description{
XXX Describe derivation of data set here and what the columns mean
}
\keyword{data}

