% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{odsSlp}
\alias{odsSlp}
\title{Slope Test Data}
\format{An object of class \code{data.frame} with 2250 rows and 8 columns.}
\usage{
odsSlp
}
\description{
XXX Describe derivation of data set here and what the columns mean
}
\keyword{data}
