% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{subject.gradient.ll.lme}
\alias{subject.gradient.ll.lme}
\title{Subject specific contribution to the lme model score (also returns marginal Vi=Cov(Y|X))}
\usage{
subject.gradient.ll.lme(yi, xi, zi, beta, sigma0, sigma1, rho, sigmae)
}
\arguments{
\item{yi}{n_i-response vector}

\item{xi}{n_i by p design matrix for fixed effects}

\item{zi}{n_i by 2 design matric for random effects (intercept and slope)}

\item{beta}{mean model parameter p-vector}

\item{sigma0}{std dev of the random intercept distribution}

\item{sigma1}{std dev of the random slope distribution}

\item{rho}{correlation between the random intercept and slope}

\item{sigmae}{std dev of the measurement error distribution}
}
\value{
Subject specific contribution to the log-likelihood score (also returns marginal Vi=Cov(Y|X))
}
\description{
Subject specific contribution to the lme model score (also returns marginal Vi=Cov(Y|X))
}
