% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R, R/Functions5.R
\name{vi.calc}
\alias{vi.calc}
\title{Calculate V_i = Z_i D t(Z_i) + sig_e^2 I_{n_i}}
\usage{
vi.calc(zi, sigma.vc, rho.vc, sigma.e)

vi.calc(zi, sigma.vc, rho.vc, sigma.e)
}
\arguments{
\item{zi}{n_i by 2 design matrix for the random effects}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}

\item{sigma0}{std dev of the random intercept distribution}

\item{sigma1}{std dev of the random slope distribution}

\item{rho}{correlation between the random intercept and slope}

\item{sigmae}{std dev of the measurement error distribution}

\item{zi}{n_i by q design matrix for the random effects}
}
\value{
V_i

V_i
}
\description{
Calculate V_i = Z_i D t(Z_i) + sig_e^2 I_{n_i}

Calculate V_i = Z_i D t(Z_i) + sig_e^2 I_{n_i}
}
