% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{acml.lmem}
\alias{acml.lmem}
\title{Fitting function: ACML for a linear mixed effects model (random intercept and slope)}
\usage{
acml.lmem(formula.fixed, formula.random, data, id, w.function = "mean",
  InitVals, cutpoints = c(0, 5), SampProb = NA, SampProbiWL = NA,
  ProfileCol = NA)
}
\arguments{
\item{formula.fixed}{formula for the fixed effects (of the form y~x)}

\item{formula.random}{formula for the random effects (of the form ~z)}

\item{data}{data frame (should contain everything in formula.fixed, formula.random, id, and SampProbiWL)}

\item{id}{sum(n_i) vector of subject ids}

\item{w.function}{options include "mean" "intercept" "slope" and "bivar"}

\item{InitVals}{starting values for c(beta, log(sigma0), log(sigma1), rho, log(sigmae))}

\item{cutpoints}{cutpoints defining the sampling regions [bivariate Q_i: a vector of length 4 c(xlow, xhigh, ylow, yhigh);
univariate Q_i: a vector of length K c(k1,k2, ... K) to define the cutpoints for Q_i based sampling regions]}

\item{SampProb}{Sampling probabilities from within each region [bivariate Q_i: a vector of length 2 c(central region, outlying region);
univariate Q_i: a vector of length K+1 with sampling probabilities for each region]}

\item{SampProbiWL}{Subject specific sampling probabilities.  A vector of length sum(n_i).  Not used unless using weighted Likelihood}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters
while fixing these columns at the values of params[ProfileCol]}
}
\value{
Ascertainment corrected Maximum likelihood: Ests, covar, LogL, code, robcov
}
\description{
Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)
}

