% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions4.R
\name{logACi2q.score2}
\alias{logACi2q.score2}
\title{Gradient of the log of the ascertainment correction piece for sampling based on bivariate Q_i}
\usage{
logACi2q.score2(subjectData, beta, sigma0, sigma1, rho, sigmae)
}
\arguments{
\item{subjectData}{a list containing: yi, xi, zi, Weights.i, w.function, SampProb, cutpoints}

\item{beta}{mean model parameter p-vector}

\item{sigma0}{std dev of the random intercept distribution}

\item{sigma1}{std dev of the random slope distribution}

\item{rho}{correlation between the random intercept and slope}

\item{sigmae}{std dev of the measurement error distribution}
}
\value{
gradient of the log transformed ascertainment correction under the bivariate sampling design
}
\description{
Calculate the gradient of the log transformed ascertainment correction under designs that sample based on a bivariate Q_i (numerically)
}
